Babel integrated with webpack
