+++ cvatjs/webpack.config.js
@@ -1,16 +1,58 @@
 /* global
     require:true,
-    __dirname:true
+    __dirname:true,
 */
 
 const path = require('path');
+const webpack = require('webpack');
+
+const configuration = 'production';
+const target = 'web';
+const plugins = [];
+
+if (configuration === 'production' && target === 'web') {
+    plugins.push(
+        /**
+         * IgnorePlugin will skip any require
+         * that matches the following regex.
+         */
+        new webpack.IgnorePlugin(/browser-env/),
+    );
+}
 
 module.exports = {
-    mode: 'development',
+    mode: configuration,
     devtool: 'source-map',
-    entry: `${path.resolve(__dirname, 'babel.build')}/api.js`,
+    entry: './src/api.js',
     output: {
         path: path.resolve(__dirname, 'dist'),
         filename: 'cvat.js',
     },
+    module: {
+        rules: [{
+            test: /.js?$/,
+            exclude: /node_modules/,
+            use: {
+                loader: 'babel-loader',
+                options: {
+                    presets: [
+                        ['@babel/preset-env', {
+                            targets: {
+                                chrome: 58,
+                            },
+                            useBuiltIns: 'usage',
+                            loose: false,
+                            spec: false,
+                            debug: false,
+                            include: [],
+                            exclude: [],
+                        }],
+                    ],
+                    sourceType: 'unambiguous',
+                },
+            },
+        }],
+    },
+    plugins,
+    target,
 };
